Bug fix in Mul()
When I converted from Radix 10^3 to Radix 10^9, a conversion between the two radix were introduced in Mul(), since the FFT still use radix  10^3 , converting back to radix 10^9  I  read beyond the FFT buffer, when it wasn't a multiple of 3. The fix overallocates and initializes  the buffers with two extra elements to avoid strange values.
Obvious, when you know it.